refactor: split the story gallery module into one file per component
