+++ src/components/blocks/ActionableList/ActionableList.stories.tsx
@@ -2,7 +2,8 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 
 import { ActionableList, type ActionableListOption } from '@blocks/ActionableList/ActionableList'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const meta: Meta<typeof ActionableList> = {
   title: 'Blocks/ActionableList',
