+++ src/components/blocks/ExpandableCard/ExpandableCard.stories.tsx
@@ -4,7 +4,8 @@ import { HStack } from '@ui/Stack/Stack'
 import { P, Span } from '@ui/Typography/Text'
 import { ExpandableCard } from '@blocks/ExpandableCard/ExpandableCard'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const noop = () => {}
