+++ src/components/blocks/FileThumb/FileThumb.stories.tsx
@@ -2,7 +2,8 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 
 import { FileThumb } from '@blocks/FileThumb/FileThumb'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const meta: Meta<typeof FileThumb> = {
   title: 'Blocks/FileThumb',
