+++ src/components/blocks/MobileSelectionDrawer/MobileSelectionDrawerWithTrigger.stories.tsx
@@ -5,7 +5,8 @@ import { userEvent, within } from 'storybook/test'
 import type { ComboBoxOption } from '@ui/ComboBox/types'
 import { MobileSelectionDrawerWithTrigger } from '@blocks/MobileSelectionDrawer/MobileSelectionDrawerWithTrigger'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const OPTIONS: ComboBoxOption[] = [
   { label: 'Profit and Loss', value: 'pnl' },
