+++ src/components/blocks/Table/PaginatedDataTable/PaginatedDataTable.stories.tsx
@@ -16,7 +16,8 @@ import {
   TableStoryGridStyles,
   TableStoryStyles,
 } from '@testUtils/storybook/data/tables'
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const COLUMN_CONFIG = getCustomerColumnConfig()
 const PINNED_COLUMN_CONFIG = getPinnedCustomerColumnConfig()
