+++ src/components/ui/Checkbox/Checkbox.stories.tsx
@@ -2,7 +2,9 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 
 import { Checkbox } from '@ui/Checkbox/Checkbox'
 
-import { Gallery, Row, Section } from '@testUtils/storybook/layout/gallery'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
+import { Row } from '@testUtils/storybook/layout/Row'
+import { Section } from '@testUtils/storybook/layout/Section'
 
 const VARIANTS = ['default', 'round', 'error', 'branded'] as const
 const SIZES = ['sm', 'md', 'lg'] as const
