+++ src/components/ui/Legend/Legend.stories.tsx
@@ -3,7 +3,8 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 import type { SeriesData } from '@ui/Chart/seriesTypes'
 import { Legend, LegendLayout } from '@ui/Legend/Legend'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const ITEMS: SeriesData[] = [
   { name: 'income', displayName: 'Income', value: 12000 },
