+++ src/components/ui/Switch/Switch.stories.tsx
@@ -2,7 +2,8 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 
 import { Switch } from '@ui/Switch/Switch'
 
-import { Gallery, Row } from '@testUtils/storybook/layout/gallery'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
+import { Row } from '@testUtils/storybook/layout/Row'
 
 const meta: Meta<typeof Switch> = {
   title: 'UI/Switch',
