+++ src/components/ui/Table/Table.stories.tsx
@@ -3,7 +3,9 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 import { Alignment } from '@internal-types/utility/table'
 import { Cell, Column, Row, Table, TableBody, TableHeader } from '@ui/Table/Table'
 
-import { Col, Gallery, Section } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
+import { Section } from '@testUtils/storybook/layout/Section'
 
 type ColumnKey = 'account' | 'debit' | 'credit'
