+++ src/components/ui/Tabs/Tabs.stories.tsx
@@ -4,7 +4,8 @@ import { ChartNoAxesColumn, CircleAlert, FileText, Receipt } from 'lucide-react'
 import { Badge, BadgeSize, BadgeVariant } from '@ui/Badge/Badge'
 import { Tabs } from '@ui/Tabs/Tabs'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const OPTIONS = [
   { label: 'Overview', value: 'overview' },
