+++ src/components/ui/Toggle/Toggle.stories.tsx
@@ -2,7 +2,8 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 
 import { Toggle } from '@ui/Toggle/Toggle'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 const OPTIONS = [
   { label: 'Day', value: 'day' },
