+++ src/components/ui/Typography/Typography.stories.tsx
@@ -3,7 +3,8 @@ import { type Meta, type StoryObj } from '@storybook/react-vite'
 import { Heading, type HeadingSize } from '@ui/Typography/Heading'
 import { P, Span } from '@ui/Typography/Text'
 
-import { Gallery, Row as BaseRow } from '@testUtils/storybook/layout/gallery'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
+import { Row as BaseRow } from '@testUtils/storybook/layout/Row'
 
 const HEADING_SIZES: HeadingSize[] = ['2xs', 'xs', 'sm', 'md', 'lg', 'xl', '2xl', '3xl']
 const TEXT_SIZES = ['2xs', 'xs', 'sm', 'md', 'lg', 'xl'] as const
