+++ src/testUtils/storybook/layout/formField.tsx
@@ -3,7 +3,8 @@ import { type PropsWithChildren, type ReactNode } from 'react'
 import { useAppForm } from '@hooks/features/forms/useForm'
 import type { CommonFormFieldProps } from '@blocks/Form/types'
 
-import { Col, Gallery } from '@testUtils/storybook/layout/gallery'
+import { Col } from '@testUtils/storybook/layout/Col'
+import { Gallery } from '@testUtils/storybook/layout/Gallery'
 
 /*
  * A real form sets these on its container: a fixed label column is what separates label from
